-made StreamDataMessage a template class; -made responder and requester using a pimpl idiom;
