+++ include/fcgi_responder/requester.h
@@ -1,29 +1,13 @@
 #pragma once
 #include <string>
-#include <sstream>
 #include <memory>
 #include <functional>
-#include <set>
 #include <map>
 #include <optional>
 #include <memory>
 
 namespace fcgi{
-class RecordReader;
-class Record;
-class Request;
-class Response;
-class MsgGetValuesResult;
-class MsgUnknownType;
-class MsgEndRequest;
-class MsgStdOut;
-class MsgStdErr;
-
-struct ResponseData
-{
-    std::string data;
-    std::string errorMsg;
-};
+class RequesterImpl;
 
 class RequestHandle{
 public:
@@ -34,18 +18,13 @@ private:
     std::weak_ptr<std::function<void()>> cancelRequestHandler_;
 };
 
-class Requester{
-    enum class ConnectionState{
-        NotConnected,
-        ConnectionInProgress,
-        Connected
-    };
+struct ResponseData
+{
+    std::string data;
+    std::string errorMsg;
+};
 
-    enum class ResponseStatus{
-        Successful,
-        Failed,
-        Cancelled
-    };
+class Requester{
 
 public:
     std::optional<RequestHandle> sendRequest(
@@ -59,61 +38,25 @@ public:
     int maximumRequestsNumber() const;
     bool isMultiplexingEnabled() const;
 
+    Requester(const Requester&) = delete;
+    Requester& operator=(const Requester&) = delete;
+
 protected:
     Requester();
     virtual ~Requester();
+    Requester(Requester&&) = default;
+    Requester& operator=(Requester&&) = default;
+
     void receiveData(const char* data, std::size_t size);
     virtual void sendData(const std::string& data) = 0;
     virtual void disconnect() = 0;
 
 private:
-    void initConnection(
-            const std::map<std::string, std::string>& params,
-            const std::string& data,
-            const std::function<void(const std::optional<ResponseData>&)>& responseHandler,
-            bool keepConnection);
-    std::optional<RequestHandle> doSendRequest(
-            const std::map<std::string, std::string>& params,
-            const std::string& data,
-            const std::function<void(const std::optional<ResponseData>&)>& responseHandler,
-            bool keepConnection);
-    void doEndRequest(uint16_t requestId, ResponseStatus responseStatus);
-    void onRecordRead(const Record& record);
-    template <typename TMsg>
-    void sendMessage(uint16_t requestId, TMsg&& msg);
-    void notifyAboutError(const std::string &errorMsg);
-    void sendRecord(const Record &record);
-    bool isRecordExpected(const Record& record);
-    void onGetValuesResult(const MsgGetValuesResult& msg);
-    void onUnknownType(uint16_t requestId, const MsgUnknownType& msg);
-    void onEndRequest(uint16_t requestId, const MsgEndRequest& msg);
-    void onStdOut(uint16_t requestId, const MsgStdOut& msg);
-    void onStdErr(uint16_t requestId, const MsgStdErr& msg);
+    RequesterImpl& impl();
+    const RequesterImpl& impl() const;
 
 private:
-    struct Config{
-        int maxConnectionsNumber = 1;
-        int maxRequestsNumber = 10;
-        bool multiplexingEnabled = true;
-    } cfg_;
-
-    struct ResponseContext{
-        std::function<void(const std::optional<ResponseData>&)> responseHandler;
-        ResponseData responseData;
-        bool keepConnection = false;
-        std::shared_ptr<std::function<void()>> cancelRequestHandler;
-    };
-
-    std::unique_ptr<RecordReader> recordReader_;
-    std::ostringstream recordStream_;
-    std::string recordBuffer_;
-    std::function<void(const std::string&)> errorInfoHandler_;
-    std::function<void()> onConnectionFail_;
-    std::function<void()> onConnectionSuccess_;
-    std::shared_ptr<std::function<void()>> connectionOpeningRequestCancelHandler_;
-    ConnectionState connectionState_ = ConnectionState::NotConnected;
-    std::set<uint16_t> requestIdPool_;
-    std::map<uint16_t, ResponseContext> responseMap_;
+    std::unique_ptr<RequesterImpl> impl_;
 };
 
 }
