+++ src/responderimpl.h
@@ -1,0 +1,77 @@
+#pragma once
+#include "recordreader.h"
+#include "requestregistry.h"
+#include "streamdatamessage.h"
+#include "types.h"
+#include <unordered_map>
+#include <functional>
+#include <memory>
+#include <sstream>
+
+namespace fcgi{
+class Request;
+class Response;
+class MsgBeginRequest;
+class MsgGetValues;
+class MsgParams;
+class Record;
+
+class ResponderImpl{
+public:
+    ResponderImpl(std::function<void(const std::string&)> sendData,
+                  std::function<void()> disconnect,
+                  std::function<void(Request&& request, Response&& response)> processRequest);
+    void receiveData(const char* data, std::size_t size);
+    void setMaximumConnectionsNumber(int value);
+    void setMaximumRequestsNumber(int value);
+    void setMultiplexingEnabled(bool state);
+    int maximumConnectionsNumber() const;
+    int maximumRequestsNumber() const;
+    bool isMultiplexingEnabled() const;
+    void setErrorInfoHandler(std::function<void(const std::string&)> errorInfoHandler);
+
+private:
+    void onRecordRead(const Record& record);
+    void onBeginRequest(uint16_t requestId, const MsgBeginRequest& msg);
+    void onGetValues(const MsgGetValues& msg);
+    void onParams(uint16_t requestId, const MsgParams& msg);
+    void onStdIn(uint16_t requestId, const StreamDataMessage<RecordType::StdIn>& msg);
+    void onRequestReceived(uint16_t requestId);
+    void sendRecord(const Record& record);
+    void sendResponse(uint16_t id, std::string&& data, std::string&& errorMsg);
+
+    bool isRecordExpected(const Record& record);
+    void endRequest(uint16_t requestId, ProtocolStatus protocolStatus);
+
+    void notifyAboutError(const std::string& errorMsg);
+    void createRequest(uint16_t requestId, bool keepConnection);
+    void deleteRequest(uint16_t requestId);
+
+private:
+    struct Config{
+        int maxConnectionsNumber = 1;
+        int maxRequestsNumber = 10;
+        bool multiplexingEnabled = true;
+    } cfg_;
+
+    struct RequestSettings
+    {
+        bool keepConnection = true;
+    };
+
+    RecordReader recordReader_;
+    RequestRegistry requestRegistry_;
+    std::unordered_map<uint16_t, RequestSettings> requestSettingsMap_;
+    std::function<void(const std::string&)> errorInfoHandler_;
+    std::ostringstream recordStream_;
+    std::string recordBuffer_;
+    std::function<void(const std::string&)> sendData_;
+    std::function<void()> disconnect_;
+    std::function<void(Request&& request, Response&& response)> processRequest_;
+
+private:
+    template<typename TMsg>
+    void sendMessage(uint16_t requestId, TMsg&& msg);
+};
+
+}
